+++ packages/web/src/controls/Checkbox.tsx
@@ -62,7 +62,7 @@ const CheckboxWithRef = forwardRef(function CheckboxWithRef<CheckboxValue extend
     controlColor = 'fgInverse',
     background = checked || indeterminate ? 'bgPrimary' : 'bg',
     borderColor = checked || indeterminate ? 'bgPrimary' : 'bgLineHeavy',
-    borderRadius,
+    borderRadius = 100,
     borderWidth = 100,
     elevation,
     ...props