+++ packages/mcp-server/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.53.0 ((3/16/2026, 01:45 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.52.2 ((3/11/2026, 10:02 AM PST))
 
 This is an artificial version bump with no new change.